+++ src/design/cards/interactiveBox/index.tsx
