+++ src/design/cards/interactiveBox/styles.ts
