+++ src/design/layout/corePageStructure/config/apps.config.tsx
