+++ src/design/layout/corePageStructure/types.ts
