+++ src/pages/selectBusinessUnits/interface.tsx
@@ -4,8 +4,7 @@ import { Grid } from "@inubekit/grid";
 import { Stack } from "@inubekit/stack";
 import { Text } from "@inubekit/text";
 import { useMediaQueries } from "@inubekit/hooks";
-
-import { AppContext } from "@context/AppContext";
+import { AppContext } from "@context/authAndDataProvider";
 import { basic } from "@design/tokens";
 import {
   StyledWelcomeContainer,
