+++ src/routes/navigationSetup/index.tsx
@@ -6,11 +6,11 @@ import {
 import { PrivilegesRoutes } from "@routes/privileges";
 import { RulesRoutes } from "@routes/rules";
 import { SelectBusinessUnitsRoutes } from "@routes/selectBusinessunits";
-
 import { ErrorPage } from "@design/layout/ErrorPage";
-import { AppPage } from "@design/layout/AppPage";
+
 import { LogOut } from "@pages/app/useApp/logOut";
 import { FirstPage } from "@pages/app/useApp/firstPage";
+import { AppPage } from "@design/layout/corePageStructure";
 
 const router = createBrowserRouter(
   createRoutesFromElements(
